Registration/requirements.txt
fastapi
psycopg2-binary
pydantic
python-dotenv
uvicorn
